+++ rs-dam/entities/entities-domain/src/main/java/fr/cnes/regards/modules/entities/domain/attribute/FloatAttribute.java
@@ -11,7 +11,7 @@ import fr.cnes.regards.modules.models.domain.attributes.AttributeType;
  * @author Marc Sordi
  *
  */
-public class FloatAttribute extends AbstractAttribute<Float> {
+public class FloatAttribute extends AbstractAttribute<Double> {
 
     @Override
     public boolean represents(AttributeType pAttributeType) {
