+++ rs-dam/entities/entities-domain/src/main/java/fr/cnes/regards/modules/entities/domain/attribute/FloatIntervalAttribute.java
@@ -12,7 +12,7 @@ import fr.cnes.regards.modules.models.domain.attributes.AttributeType;
  * @author Marc Sordi
  *
  */
-public class FloatIntervalAttribute extends AbstractAttribute<Interval<Float>> {
+public class FloatIntervalAttribute extends AbstractAttribute<Interval<Double>> {
 
     @Override
     public boolean represents(AttributeType pAttributeType) {
